requirements.txt
PySide6
shiboken6
pyinstaller
ssh2-python